docs: split op_server examples and document SqlxOpStore migrate (#124)
* docs: update version references to 0.2.3

* docs: reorganize data and storage section to clarify sql store vs kv store

* docs: split op_server examples and document SqlxOpStore migrate

* style: fix rustfmt formatting in op_server examples

* fix: add RedisStore::with_client, instrument CompositeOpStore, fix multi-connection examples

* fix: correct actix op_server HttpServer wiring (op_actix_scope not op_actix_config)

* docs: clarify custom grant example comment — framework already checks client authorization
